ci: test wownero regressions as well

+++ tests/regression.test.ts
@@ -1,15 +1,20 @@
-import { $, createWalletViaCli, downloadMoneroCli, getMoneroC, getMoneroCTags } from "./utils.ts";
+import { $, createWalletViaCli, downloadCli, getMoneroC, getMoneroCTags } from "./utils.ts";
 
-Deno.test("Regression tests", async (t) => {
+const coin = Deno.env.get("COIN");
+if (coin !== "monero" && coin !== "wownero") {
+  throw new Error("COIN env var invalid or missing");
+}
+
+Deno.test(`Regression tests (${coin})`, async (t) => {
   await Deno.remove("./tests/wallets", { recursive: true }).catch(() => {});
   await Deno.mkdir("./tests/wallets", { recursive: true });
 
   const tags = await getMoneroCTags();
   const latestTag = tags[0];
-  await Promise.all([getMoneroC("next"), await getMoneroC(latestTag), downloadMoneroCli()]);
+  await Promise.all([getMoneroC(coin, "next"), await getMoneroC(coin, latestTag), downloadCli(coin)]);
 
   await t.step("Simple (next, latest, next)", async () => {
-    const walletInfo = await createWalletViaCli("dog", "sobaka");
+    const walletInfo = await createWalletViaCli(coin, "dog", "sobaka");
 
     for (const version of ["next", latestTag, "next"]) {
       await $`deno run -A ./tests/compare.ts ${version} ${JSON.stringify(walletInfo)}`;
@@ -19,10 +24,10 @@ Deno.test("Regression tests", async (t) => {
   await t.step("All releases sequentially (all tags in the release order, next)", async () => {
     tags.unshift("next");
 
-    const walletInfo = await createWalletViaCli("cat", "koshka");
+    const walletInfo = await createWalletViaCli(coin, "cat", "koshka");
 
     for (const version of tags.toReversed()) {
-      if (version !== "next" && version !== tags[0]) await getMoneroC(version);
+      if (version !== "next" && version !== tags[0]) await getMoneroC(coin, version);
       await $`deno run -A ./tests/compare.ts ${version} ${JSON.stringify(walletInfo)}`;
     }
